Refactor: Rectangle 내부의 구조체 타입들 모두 외부로 이동
