+++ DrawingApp/DrawingApp/Model/RectangleApplicable.swift
@@ -1,5 +1,8 @@
 import Foundation
 
 protocol RectangleApplicable{
-    
+    var id: Id {get}
+    var size: Size {get}
+    var point: Point {get}
+    var alpha: Alpha {get}
 }
